Refactor all utils to single file and db_utils to utilize framework db manager
