--- src/wazuh_testing/modules/aws/exceptions.py
@@ -1,2 +1,0 @@
-class OutputAnalysisError(Exception):
-    pass
